* begin __CRAZY cleanup of config(8), not yet completed   due to time constraints * store the kernel config as section .config.gz, gzip'd

+++ src/usr.sbin/config/mkheaders.c
@@ -148,7 +148,7 @@ emitopt(struct nvlist *nv)
 			    option->nv_name);
 	}
 
-	if (totlen < 0 || totlen >= sizeof new_contents) {
+	if (totlen < 0 || (size_t)totlen >= sizeof new_contents) {
 		fprintf(stderr, "config: string too long\n");
 		return (1);
 	}
